\name{listOrNULL-class}
\Rdversion{1.1}
\docType{class}
\alias{listOrNULL-class}

\title{Class \code{"listOrNULL"}}
\description{
This is a class union to represent a \code{list} or \code{NULL}. It is used for the dimnames of \code{h5array} and \code{h5matrix} objects, which can be a \code{list} of the same length as the dimensions themselves, or \code{NULL}.
}
\section{Objects from the Class}{A virtual Class: No objects may be created from it.}
\section{Methods}{
No methods defined with class "listOrNULL" in the signature.
}
\references{
%%  ~~put references to the literature/web site here~~
}
\author{
%%  ~~who you are~~
}
\note{
%%  ~~further notes~~
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
%%  ~~objects to See Also as \code{\link{~~fun~~}}, ~~~
%%  ~~or \code{\linkS4class{CLASSNAME}} for links to other classes ~~~
}
\examples{
showClass("listOrNULL")
}
\keyword{classes}
